requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
faiss-cpu
openai
python-dotenv
jinja2
aiofiles
python-multipart
passlib[bcrypt]
python-jose[cryptography]
slowapi
sentry-sdk
sqlalchemy
pydantic
pytest
httpx
email-validator
psycopg2-binary

 
 